fix root path error

+++ unix/replace-my-vim.sh
@@ -32,9 +32,16 @@ then
 	#git reset --hard
 	#patch -p1 < 0001-Segments-Add-Segments.patch
 
-	sed -i "s/\$WHOAMI/$WHOAMI/" segments/git_ext.py
-	sed -i "s/\$WHOAMI/$WHOAMI/" segments/android_prj.py
-	sed -i "s/\$WHOAMI/$WHOAMI/" segments/git_branch.py
+    if [ "$WHOAMI" == "root" ]
+    then
+	    sed -i "s/\/home\/\$WHOAMI/\/$WHOAMI/" segments/git_ext.py
+	    sed -i "s/\/home\/\$WHOAMI/\/$WHOAMI/" segments/android_prj.py
+	    sed -i "s/\/home\/\$WHOAMI/\/$WHOAMI/" segments/git_branch.py
+    else
+	    sed -i "s/\/home\/\$WHOAMI/\/home\/$WHOAMI/" segments/git_ext.py
+	    sed -i "s/\/home\/\$WHOAMI/\/home\/$WHOAMI/" segments/android_prj.py
+	    sed -i "s/\/home\/\$WHOAMI/\/home\/$WHOAMI/" segments/git_branch.py
+    fi
 	chmod a+x segments/get_git_info.sh
 	chmod a+x segments/get_android_prj.sh
 	chmod a+x segments/get_git_branch.sh
@@ -45,9 +52,20 @@ cd ${ORIGINAL_PATH}
 cp vimfiles/bundle/vim_plugins/a.vim/plugin/ -r vimfiles/
 
 rm -rf $HOME/.exvim
-mv $HOME/.vim $HOME/.vim_bk
-mv $HOME/.vimrc $HOME/.vimrc_bk
-rm -rf $HOME/.exvim*
+if [ -f $HOME/.vim ]
+then
+	mv $HOME/.vim $HOME/.vim_bk
+fi
+
+if [ -f $HOME/.vimrc ]
+then
+	mv $HOME/.vimrc $HOME/.vimrc_bk
+fi
+
+if [ -d $HOME/.exvim ]
+then
+	rm -rf $HOME/.exvim*
+fi
 
 cp ./dist/ctags_lang      $HOME/.ctags
 
@@ -61,11 +79,16 @@ ln -s $HOME/.exvim $HOME/.vim
 # cd $ORIGINAL_PATH
 
 cp .tmux.conf ~/
-sed -i "s/\$WHOAMI/$WHOAMI/" ~/.tmux.conf
 cp .bashrc ~/
 cp .zshrc ~/
-sed -i "s/\$WHOAMI/$WHOAMI/" ~/.zshrc
 cp .bashrc ~/
-
+if [ "$WHOAMI" == "root" ]
+then
+    sed -i "s/\/home\/\$WHOAMI/\/$WHOAMI/" ~/.tmux.conf
+    sed -i "s/\/home\/\$WHOAMI/\/$WHOAMI/" ~/.zshrc
+else
+    sed -i "s/\/home\/\$WHOAMI/\/home\/$WHOAMI/" ~/.tmux.conf
+    sed -i "s/\/home\/\$WHOAMI/\/home\/$WHOAMI/" ~/.zshrc
+fi
 echo "install DONE!"
